one-off/pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "mokelumne-concept"
version = "0.0.1"
dependencies = [
    "langchain[aws]",
    "python-tind-client @ git+https://github.com/BerkeleyLibrary/python-tind-client@main",
    "rich ~= 14.1",
]
requires-python = ">= 3.12.12"
authors = [
    {name = "Anna Wilcox"},
]
readme = "README.md"
license = "MIT"

[project.optional-dependencies]
test = [
]
lint = [
    "mypy ~= 1.17.1",
    "pydoclint ~= 0.6.10",
    "pylint ~= 3.3",
]

[project.urls]
Repository = "https://github.com/BerkeleyLibrary/mokelumne"
Issues = "https://github.com/BerkeleyLibrary/mokelumne/issues"

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_redundant_casts = true
warn_return_any = true

[tool.pydoclint]
allow-init-docstring = true
skip-checking-raises = true
style = "sphinx"
